Start of a mammoth refactoring.
The goal of the refactoring is to eliminate circular dependencies among
core packages, and among interfaces of the core package. In this first
step, the com.choicemaker.cm.core package was renamed to
com.choicemaker.cm.core.base. It will eventually contain only implementation
classes. At this stage, there is no com.choicemaker.cm.core package, but
in later refactorings this package will be recreated and all core
interfaces will be moved to it (along with a small number of concrete
classes).

The code still compiles with this commit; CM Analyzer still works; and
ClueMaker files still compile to working Java classes.

A huge number of projects, packages, classes, scripts and configuration
files are changed with this commit.