Continuation of a mammoth refactoring.
See previous commit for overview. This step of the refactoring
recreated the com.choicemaker.cm.core package and moved most interfaces,
Exceptions, and some enum-like concrete classes into the core pacakge.

The code still compiles with this commit; CM Analyzer still works; but
ClueMaker files no longer compile to working Java classes. The ClueMaker
issue will be fixed in the next commit.

As with the previous commit, a huge number of projects, packages, classes,
scripts and configuration files are changed with this commit.