Continuation of a mammoth refactoring (third in a series)
See previous two commits for overview. This step of the refactoring
fixes problems in the ClueMaker compiler and various generator plugins
that were preventing ClueMaker file from compiling into valid Java
code.

The code compiles with this commit; CM Analyzer works; and
ClueMaker files compile to working Java classes.

In contrast with the previous commits, a relatively modest
number of projects, packages, classes, scripts and configuration
files are changed with this commit.