Continuation of a mammoth refactoring (fourth in a series)
See previous commits for overview. This step of the refactoring
creates a new package, com.choicemaker.util, for general utilities
that have no dependency on other ChoiceMaker software. It also
renames an existing package, com.choicemaker.cm.core.configure, to
com.choicemaker.cm.configure.xml, and then reuses the 'configure'
package name for new interfaces for replacing the soon-to-be deprecated
classes XmlConfigurator, ProbabilityModelXmlConf and MachineLearnerXmlConf.
Some new interfaces are added to the core package. Not all of these new
interfaces are implemented right now, but they will be in future commits.

Moved MachineLearner, OperationFailedException and XmlConfException to core package.

Replaced ChainedException and ChainedIOException with Exception and IOException,
respectively.

New interfaces:
  Firings
  ModelTrainer
  ModelTrainingException
  ModelTrainingRegistry
  PairEvaluation
  PairEvaluator
  ProbabilityModelSpecification
  PropertyNames
  WellKnownPropertyValues

  configure/ChoiceMakerConfiguration
  configure/ChoiceMakerConfigurator
  configure/MachineLearnerPersistence
  configure/ProbabilityModelConfiguration
  configure/ProbabilityModelPersistence

  install/InstallableChoiceMakerConfigurator
  install/InstallableCompiler
  install/InstalledChoiceMakerConfiguration