First "working" prototype of the ClueMaker Maven Plugin.
Prototype is named MyMojo2 in the misc/mymojo-plugin. It "works" from the
command line when processing the POM file in the Integration Testing (it)
set of resources under the src/test directory. See below for caveats
about "working".

Also cleaned up some manifest constants in CompilerArguments and
SystemPropertyUtils.

Caveats: the prototype only works with a small change to the GeneratorImpl
class in the compiler project. This change is not commited here, but
will be commited next as a separate one-line, one-final change. The
change breaks CM Analyzer and the Compiler Application, and so it needs
to be tracked by itself. It should be straightforward to fix, although
it will have ripple effects beyond the GeneratorImpl class.