Changed the compiler fork option to false in mymojo-plugin-it.
The com.choicemaker.cm.core code won't compile under Java 1.8, but
will compile under Java 1.7. On Linux platforms, it is straight-
forward to set the JAVA_HOME environment variable to a 1.7 JDK. On
Mac platforms, one has to use the /usr/libexec/java_home tool:

  export JAVA_HOME=`/usr/libexec/java_home -v 1.7`

However, this doesn't seem to be sufficient to get the
mymojo-plugin-it project to pass the integration tests.
It seems that the compiler fork option has to be set to false
in both the project POM and the test POM's (e.g.
src/test/resources/smoke-test/pom.xml).