Testability improvements in ModelMaker.java and ModelMakerTest.java
Refactored two new methods from the ModelMaker.run(Object) method,
startup(Object) and tearDown(). The responsibilities of the run(Object)
method are simplified to invoking startup(Object) and tearDown(), and
for displaying the GUI and waiting for user input. The default
sequence is:
 (*) startup(Object) -- prepares but does not display the GUI
 (*) run(Object) -- displays the GUI and waits for user input
 (*) tearDown() -- hides the GUI and disposes of it
For headless testing, the startup(Object) and tearDown() methods
can be used by themselves, without invoking run(Object).

The integration test for ModelMaker is greatly simplified. All
the threading logic, hacks and workarounds have been removed. The
test now works reliably in a display environment, and should soon
work reliably in headless environments as well (although it has
not yet been tested this way).