Rewrote cm-server-ejb to use cm-server-plugin-discovery
Moved PluginDiscovery interface and implementation to the
cm-server-plugin-discovery project. Renamed the EJB interface
to PluginDiscoveryEJB and moved it to the fake/ejb package.
Similarly renamed the EJB implmentation to PluginDiscoverySingletonBean
and moved it to the fake/ejb package. Renamed the integration
test to PluginDiscoverySingletonBeanTest. Copied the META-INF
test resources from the cm-server-plugin-discovery project and
copied most of the logic from that project's unit test into
PluginDiscoverySingletonBeanTest.

All test configurations work, except for the arquillian-glassfish-embedded
test, which fails because the cm-server-plugin-discovery classes require
log4j classes and these are not part of the EJB test deployment.
(The glassfish-remote test works because log4j was added to the
server configuration; see http://links.rph.cx/1rXD9T0 and
http://links.rph.cx/1rXDgy3 for detailed instructions.)

Rather than modify the EJB test deployment to include log4j
classes, another test will be constructed to create an EAR file
with log4j as library, along with the META-INF resources in a
separate library file.