Created EAR test.
Refactored getExpected() [plug ids] from existing
PluginDiscoverySingletonTest into new Utils class for re-use
by EarTest. Added Utils.class to EJB deployment of
PluginDiscoverySingletonTest.

Changed version from 1.0.3.Final to 1.1.2.Final for
dependency on arquillian-bom. Added dependencies on
shrinkwrap-resolver-impl for EarTest.

EarTest and PluginDiscoverySingletonTest work in
jbossas managed and remote configurations glassfish
remote configurations.

Both tests fail in the glassfish embedded configuration
for the same reason, the expected number of plugin ids
is diffferent than the discovered number. The problem
is that the container JVM is the same ("embedded", duh) as
the JVM in which the test is built, and therefore two
copies of the META-INF/plugins/* entries are discovered.