automated.offline.server: Initial, incomplete version of CM_ModelBean
CM_ModelBean persists information about ChoiceMaker probability model to
a database. The persisted information consists of signatures for the
model schema, clue set, and weights, plus the model name, clue set
name, and schema name. This information is not sufficient to recreate a
model, but it is sufficient to verify to a reasonable degree of
certainty whether a model has changed over time.

Added a dependency on Apache commons-collections4 since the CM_ModelBean
is implemented with a bi-directional map.

Added a simple integration test to check that model signatures and names
can be saved to and loaded from a database. The test uses a stubbed
version of an ImmutableProbabilityModel, so it runs independently
of any particular model (and of the Eclipse2 plugin framework).

The CM_ModelBean class is still incomplete. It still requires lists
of associated database and blocking configurations, plus a list of
a model's associated clues and weights.

Modified the cm.core.ImmutableProbabilityModel interface to add accessors:
* getClueSetName
* getClueSetSignature
* getModelSignature
* getSchemaName
* getSchemaSignature
Implemented corresponding methods in cm.core.base.MutableProbabilityModel.

Modified the abstract cm.core.Evaluator class to add an abstract
'getSignature()' method that subclasses must define.

Implemented the getSignature() methdod in the Maximum Entropy evaluator
class (MeEvaluator).

Added a cm.util.Signature class to standardize signature calculations
for various types of objects.

Modified the deprecated IProbabilityModel interface to remove modifiers
* setModelName
* setClueSetName
These methods remain on the deprecated ProbabilityModel class, but they should
not be on the interface.

Modified the ProbabilityModelXmlConf class so that it sets a model name
using a method on the MutableProbabilityModel class (because the setModelName
method is removed from the IMutableProbabilityModel interface).

Removed commented-out code from cm.core.MachineLearner