cmit: Another hack in DedupOabaIT
Defined CallableMessage<T> and subclassed
CallableOabaJobMessage as CallableMessage<OabaJobMessage>.
Used this call to get a message that is expected to be significantly
delayed. The CallMessage... classes are not hacks by themselves, but
the way that they are being used in DedupOABA is a hack, since it
amounts to little more than calling Thread.sleep(..) inside an EJB
container. The proper way to do this would be to use yet another MDB,
possibly paired with an EJB stateless or singleton bean, or possibly
to use EJB managed threads. However, this is **TEST** code, so if this
hack works, it is good enough for now.