transitivity: Redefined TransitivityJob as extending BatchJob
Previously, TransitivityJob extended OabaJob, but this was confusing
since a transitivity job has parameters and settings that an OABA job
does not. The BatchJob interface is minimal, and does not define
parameters or settings, so there's no confusion about which methods
are relevant.

Removed deprecated methods from the TransitivityJob interface.

Defined constructors for TransitityJobEntity that are analogous to
constructors for OabaJobEntity. Most of TransitivityJobEntity remains
stubbed.

Moved methods used only by the integration test for TransitivityJobEntity
from EntityManagerUtils to TransitivityJobEntityIT.