oaba: Deprecated RecordIdTranslator2
This implementation can fail because it assumes that the same number of
record ids comes back from the database each time and in exactly the same
order. However, the OABA currently pulls records from the database three times
(during OABA blocking and chunking, and then during transitivity analysis).
The database can change between these pulls, and in some installations does,
which changes the number and ordering of the record ids that are retrieved.

Reordered methods alphabetically, to facilitate comparison with RecordIdTranslator3.